+++ src/loader.js
@@ -2,6 +2,7 @@ const { Client, Collection, Routes } = require('discord.js');
 const fs = require('fs');
 const { info, warn, error, nolog } = require('./log.js');
 const { REST } = require('@discordjs/rest');
+const { betacommands, betaserver } = require('../config.json');
 
 /**
  * @param {Client} client
@@ -10,6 +11,8 @@ const { REST } = require('@discordjs/rest');
 function load(client) {
     // Initialize collections for commands, buttons, and modals
     client.commands = new Collection();
+    client.globalcommands = new Collection();
+    client.betacommands = new Collection();
     client.buttons = new Collection();
     client.modals = new Collection();
 
@@ -17,14 +20,29 @@ function load(client) {
     const commandFiles = fs.readdirSync(`./Commands`);
 
     for (const file of commandFiles) {
-        const filePath = `../Commands/${file}`;
-        const command = require(filePath);
-        // Set a new item in the Collection with the key as the command name and the value as the exported module
-        if ('data' in command && 'execute' in command) {
-            client.commands.set(command.data.name, command);
+        if(betacommands.includes(file)) {
+            const filePath = `../Commands/${file}`;
+            const command = require(filePath);
+            // Set a new item in the Collection with the key as the command name and the value as the exported module
+            if ('data' in command && 'execute' in command) {
+                client.betacommands.set(command.data.name, command);
+                client.commands.set(command.data.name, command);
+            } else {
+                warn(`The command at ${filePath} is missing a required "data" or "execute" property.`);
+            }           
         } else {
-            warn(`The command at ${filePath} is missing a required "data" or "execute" property.`);
+            const filePath = `../Commands/${file}`;
+            const command = require(filePath);
+            // Set a new item in the Collection with the key as the command name and the value as the exported module
+            if ('data' in command && 'execute' in command) {
+                client.globalcommands.set(command.data.name, command);
+                client.commands.set(command.data.name, command);
+            } else {
+                warn(`The command at ${filePath} is missing a required "data" or "execute" property.`);
+            }
         }
+
+
     }
 
     // Read all Modals
@@ -42,7 +60,7 @@ function load(client) {
     }
 
     // Log the number of commands, buttons, and modals that were loaded
-    info(`Loaded ${client.commands.size} commands, ${client.buttons.size} buttons, ${client.modals.size} modals`);
+    info(`Loaded ${client.globalcommands.size} commands, ${client.buttons.size} buttons, ${client.modals.size} modals, ${client.betacommands.size} beta commands.`);
 }
 
 /**
@@ -56,12 +74,21 @@ async function register(client) {
         info(`Started refreshing ${client.commands.size} application (/) commands.`);
         
         // Create an array of commands
-        const commands = client.commands.map(command => command.data.toJSON());
+        const commands = client.globalcommands.map(command => command.data.toJSON());
 
         // Use the REST API to register commands
         await rest.put(
             Routes.applicationCommands(client.user.id),
             { body: commands }
+        );
+
+        // Create an array of commands
+        const betacommandse = client.betacommands.map(command => command.data.toJSON());
+
+        // Use the REST API to register commands
+        await rest.put(
+            Routes.applicationGuildCommands(client.user.id, betaserver),
+            { body: betacommandse }
         );
 
         info(`Successfully reloaded ${client.commands.size} application (/) commands.`);